t-18 all interaction moved into single handler
